requirements.txt
fonttools==4.60.1
Pillow==12.0.0
Requests==2.32.5
